extract wait90TopviewThread from robot_lib
