+++ robot_lib.py
@@ -19,6 +19,7 @@ import filecmp
 import _thread
 import logging
 import epics.ca
+import top_view
 
 from config_params import TOP_VIEW_CHECK
 logger = logging.getLogger(__name__)
@@ -50,78 +51,7 @@ def warmupGripperRecoverThread(savedThreshold,junk):
   time.sleep(120.0)
   setPvDesc("warmupThreshold",savedThreshold)      
 
-def wait90TopviewThread(prefix1,prefix90):
-  global sampXadjust, sampYadjust, sampZadjust
-
-  sampXadjust = 0
-  sampYadjust = 0
-  sampZadjust = 0
-  startTime = time.time()
-  if (getPvDesc("gripTemp")>-170):
-    threadTimeout = 130.0
-  else:
-    threadTimeout = 30.0    
-  while (1):
-    omegaCP = beamline_lib.motorPosFromDescriptor("omega")
-    if (omegaCP > 89.5 and omegaCP < 90.5):
-      logger.info("leaving topview thread for 90")      
-      break
-    if (time.time()-startTime > threadTimeout):
-      logger.info("leaving topview thread for " + str(threadTimeout))
-      setPvDesc("topViewTrigMode",0)    
-      setPvDesc("topViewImMode",2)
-      setPvDesc("topViewDataType",1)
-      setPvDesc("topViewAcquire",1,wait=False)    
-      return
-    time.sleep(0.10)
-  try:
-    daq_macros.topViewWrite()
-    daq_macros.topViewSnap(prefix90,os.getcwd()+"/pinAlign",1)
-    snapshot1Name = prefix1+"_001.jpg"
-    snapshot2Name = prefix90+"_001.jpg"
-    if (not filecmp.cmp(os.getcwd()+"/pinAlign/"+snapshot1Name,os.getcwd()+"/pinAlign/"+snapshot2Name)): #this would mean something is wrong if true because the pictures are identical
-      comm_s = os.environ["LSDCHOME"] + "/runPinAlign.py " + snapshot1Name + " " + snapshot2Name
-      logger.info(comm_s)
-      lines = os.popen(comm_s).readlines()
-      logger.info("printing lines right after popen ")
-      logger.info(lines)
-      logger.info(" done")
-      if (lines[0].find("CANNOT CENTER") == -1):
-        offsetTokens = lines[0].split()
-        logger.info(offsetTokens[0] + " " + offsetTokens[1] + " " + offsetTokens[2])
-        xlimLow = getPvDesc("robotXMountPos") + getPvDesc("robotXMountLowLim")
-        xlimHi = getPvDesc("robotXMountPos") + getPvDesc("robotXMountHiLim")
-        xpos = beamline_lib.motorPosFromDescriptor("sampleX")          
-        target = xpos + float(offsetTokens[0])*1000.0
-        if (target<xlimLow or target>xlimHi):
-          logger.info("Pin X move beyond limit - Mount next sample.")
-##else it thinks it worked              return 0
-        else:
-          sampXadjust = 1000.0*float(offsetTokens[0])
-          sampYadjust = 1000.0*float(offsetTokens[1])
-          sampZadjust = 1000.0*float(offsetTokens[2])
-          sampXCP = beamline_lib.motorPosFromDescriptor("sampleX")
-          sampYCP = beamline_lib.motorPosFromDescriptor("sampleY")
-          sampZCP = beamline_lib.motorPosFromDescriptor("sampleZ")
-          sampXAbsolute = sampXCP+sampXadjust          
-          sampYAbsolute = sampYCP+sampYadjust          
-          sampZAbsolute = sampZCP+sampZadjust
-          setPvDesc("robotXWorkPos",sampXAbsolute)
-          setPvDesc("robotYWorkPos",sampYAbsolute)
-          setPvDesc("robotZWorkPos",sampZAbsolute)          
-      else:
-        logger.info("Cannot align pin - Mount next sample.")
-#else it thinks it worked            return 0
-      for outputline in lines:
-        logger.info(outputline)
-  except Exception as e:
-    e_s = str(e)
-    message = "TopView check ERROR, will continue: " + e_s
-    daq_lib.gui_message(message)
-    logger.error(message)
-    
-  
-  
+ 
 def recoverRobot():
   try:
     rebootEMBL()
@@ -307,7 +237,7 @@ def mountRobotSample(puckPos,pinPos,sampID,init=0,warmup=0):
           if (omegaCP > 89.5 and omegaCP < 90.5):
             beamline_lib.mvrDescriptor("omega", 85.0)
           logger.info("calling thread")                        
-          _thread.start_new_thread(wait90TopviewThread,(prefix1,prefix90))
+          _thread.start_new_thread(top_view.wait90TopviewThread,(prefix1,prefix90))
           logger.info("called thread")
         setPvDesc("boostSelect",0)                    
         if (getPvDesc("gripTemp")>-170):
@@ -334,7 +264,7 @@ def mountRobotSample(puckPos,pinPos,sampID,init=0,warmup=0):
           if (omegaCP > 89.5 and omegaCP < 90.5):
             beamline_lib.mvrDescriptor("omega", 85.0)
           logger.info("calling thread")            
-          _thread.start_new_thread(wait90TopviewThread,(prefix1,prefix90))
+          _thread.start_new_thread(top_view.wait90TopviewThread,(prefix1,prefix90))
           logger.info("called thread")
         if (warmup):
           RobotControlLib._mount(absPos,warmup=True)
